+++ core-ng-test/src/test/java/core/framework/test/TestModule.java
@@ -71,8 +71,7 @@ public class TestModule extends AbstractTestModule {
     private void configureCache() {
         cache().redis("localhost");
         cache().maxLocalSize(5000);
-        cache().remote(TestDBEntity.class, Duration.ofHours(6));
-        cache().local(TestMessage.class, Duration.ofHours(6));
+        cache().add(TestDBEntity.class, Duration.ofHours(6));
     }
 
     private void configureExecutor() {