+++ CHANGELOG.md
@@ -9,7 +9,10 @@
     the official one moves slowly, https://github.com/RedisLabs/redis-cluster-proxy
 * kafka: update to 2.6.0
     high cpu issue fixed, https://issues.apache.org/jira/browse/KAFKA-10134
-
+* cache: cache config api changed 
+    !!! replace cache().local() cache().remote() back to cache().add()
+    !!! for multiple level caching, use cache().add().local()
+    
 ### 7.5.1 (07/23/2020 - 07/31/2020)
 * utils: add Sets.newEnumSet and Maps.newEnumMap shortcuts
 * httpClient: remove HTTPClientBuilder.tlsVersions(), jdk 14.0.2 fixed TLSv1.3 issues